pyproject.toml
[tool.poetry]
name = "python_basic_diploma"
version = "0.1.0"
description = ""
authors = ["jjuchara <jjuchara@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.5.3"
python-decouple = "^3.4"
aiogram = "^2.15"
uvloop = "^0.16.0"
ujson = "^4.2.0"
cchardet = "^2.1.7"
aiodns = "^3.0.0"
requests = "^2.26.0"
aiogram-calendar = "^0.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
